Move/Rename HuffmanTable as HuffmanTableLUT implementation
